Integral control added - FB control still needs tuning

+++ code_LL2/traj_ctrl_params_nml.h
@@ -59,7 +59,7 @@ traj_ellipse_help(	double phi, double dt_phi, double p[], double dt_p[], double 
 
 void
 traj_linear_points(	double p[], double dt_p[], double u_t[], double dt_k,
-					double p_o[], double p_f[], double v_max, double alpha, int* initial);
+					double p_o[], double p_f[], double v_max, double alpha, uint8_t* initial);
 
 void
 pos_linear_relative(double* dist_rel, double* v_rel, double t, double dist_max, double v_max, double alpha);
